Moving some more things to the base router class.
